+++ Debug/application/main.d
@@ -1,9 +1,14 @@
 application/main.o: ../application/main.c \
  C:\work\TryKernel\include/trykernel.h \
  C:\work\TryKernel\include/typedef.h C:\work\TryKernel\include/sysdef.h \
- C:\work\TryKernel\include/syslib.h C:\work\TryKernel\include/knldef.h
+ C:\work\TryKernel\include/syslib.h C:\work\TryKernel\include/config.h \
+ C:\work\TryKernel\include/error.h C:\work\TryKernel\include/knldef.h \
+ C:\work\TryKernel\include/apidef.h
 C:\work\TryKernel\include/trykernel.h:
 C:\work\TryKernel\include/typedef.h:
 C:\work\TryKernel\include/sysdef.h:
 C:\work\TryKernel\include/syslib.h:
+C:\work\TryKernel\include/config.h:
+C:\work\TryKernel\include/error.h:
 C:\work\TryKernel\include/knldef.h:
+C:\work\TryKernel\include/apidef.h:
