+++ Debug/kernel/inittsk.d
@@ -1,4 +1,5 @@
-kernel/syslib.o: ../kernel/syslib.c C:\work\TryKernel\include/trykernel.h \
+kernel/inittsk.o: ../kernel/inittsk.c \
+ C:\work\TryKernel\include/trykernel.h \
  C:\work\TryKernel\include/typedef.h C:\work\TryKernel\include/sysdef.h \
  C:\work\TryKernel\include/syslib.h C:\work\TryKernel\include/config.h \
  C:\work\TryKernel\include/error.h C:\work\TryKernel\include/knldef.h \
