+++ Debug/kernel/subdir.mk
@@ -5,7 +5,11 @@
 # Add inputs and outputs from these tool invocations to the build variables 
 C_SRCS += \
 ../kernel/context.c \
-../kernel/syslib.c 
+../kernel/inittsk.c \
+../kernel/scheduler.c \
+../kernel/syslib.c \
+../kernel/task_manage.c \
+../kernel/task_queue.c 
 
 S_UPPER_SRCS += \
 ../kernel/dispatch.S 
@@ -13,14 +17,22 @@ S_UPPER_SRCS += \
 OBJS += \
 ./kernel/context.o \
 ./kernel/dispatch.o \
-./kernel/syslib.o 
+./kernel/inittsk.o \
+./kernel/scheduler.o \
+./kernel/syslib.o \
+./kernel/task_manage.o \
+./kernel/task_queue.o 
 
 S_UPPER_DEPS += \
 ./kernel/dispatch.d 
 
 C_DEPS += \
 ./kernel/context.d \
-./kernel/syslib.d 
+./kernel/inittsk.d \
+./kernel/scheduler.d \
+./kernel/syslib.d \
+./kernel/task_manage.d \
+./kernel/task_queue.d 
 
 
 # Each subdirectory must supply rules for building sources it contributes
