+++ Debug/kernel/task_manage.d
@@ -1,4 +1,5 @@
-kernel/syslib.o: ../kernel/syslib.c C:\work\TryKernel\include/trykernel.h \
+kernel/task_manage.o: ../kernel/task_manage.c \
+ C:\work\TryKernel\include/trykernel.h \
  C:\work\TryKernel\include/typedef.h C:\work\TryKernel\include/sysdef.h \
  C:\work\TryKernel\include/syslib.h C:\work\TryKernel\include/config.h \
  C:\work\TryKernel\include/error.h C:\work\TryKernel\include/knldef.h \
