+++ application/main.c
@@ -1,79 +1,63 @@
-﻿#include <trykernel.h>
+﻿/*
+ *** main.c
+ */
 
-/* Task control data */
-UINT cur_task 	= 0;		// current task ID
-UINT next_task 	= 0;		// next task ID
+#include <trykernel.h>
 
-#define MAX_FNC_ID 2		// number of tasks
-void *ctx_tbl[MAX_FNC_ID];	// pointer to the registered context data
+UW	tskstk_1[1024/sizeof(UW)];	// Stack for Task1
+ID	tskid_1;
 
-/* Stack info */
-#define STACK_SIZE 1024
-UW stack_1[STACK_SIZE/sizeof(UW)];
-UW stack_2[STACK_SIZE/sizeof(UW)];
+/* Task1 create info */
+void task_1(INT stacd, void *exinf);
+T_CTSK ctsk_1 = {
+		.tskatr		= TA_HLNG | TA_RNG3 | TA_USERBUF,
+		.task		= task_1,
+		.itskpri	= 10,
+		.stksz		= 1024,
+		.bufptr		= tskstk_1,
+};
 
+UW	tskstk_2[1024/sizeof(UW)];	// Stack for Task2
+ID	tskid_2;
 
-/* Delay function */
-static void delay_ms(UINT ms)
+/* Task2 create info */
+void task_2(INT stacd, void *exinf);
+T_CTSK ctsk_2 = {
+		.tskatr		= TA_HLNG | TA_RNG3 | TA_USERBUF,
+		.task		= task_2,
+		.itskpri	= 10,
+		.stksz		= 1024,
+		.bufptr		= tskstk_2,
+};
+
+/* Task1 function */
+void task_1(INT stacd, void *exinf)
 {
-    UINT cnt = ms / TIMER_PERIOD;
-
-    while(cnt){
-        if((in_w(SYST_CSR) & SYST_CSR_COUNTFLAG) != 0){
-            cnt--;
-        }
-    }
+	tm_putstring("Start Task1!!\n");
+	tk_ext_tsk();
 }
 
-/* Task1 */
-void task_1(void)
+/* Task2 function */
+void task_2(INT stacd, void *exinf)
 {
-	while(1){
-		out_w(GPIO_OUT_SET, (1<<25));	// Turn on LED light
-		delay_ms(500);					// Wait 500msec
-
-		next_task = 2;
-		dispatch();
-	}
+	tm_putstring("Start Task2!!\n");
+	tk_ext_tsk();
 }
 
-/* Task2 */
-void task_2(void)
+int usermain(void)
 {
-	while(1){
-		out_w(GPIO_OUT_CLR, (1<<25));	// Turn off LED light
-		delay_ms(500);					// Wait 500msec
+	tm_putstring("Start usermain!!\n");
 
-		next_task = 1;
-		dispatch();
-	}
-}
+	/* Task1 */
+	tskid_1 = tk_cre_tsk(&ctsk_1);
+	tk_sta_tsk(tskid_1, 0);
 
-/* Main function */
-int main(void)
-{
-	/* Initialize task */
-	ctx_tbl[0] = make_context(stack_1, sizeof(stack_1), task_1);
-	ctx_tbl[1] = make_context(stack_2, sizeof(stack_2), task_2);
-
-	next_task = 1;
-	dispatch();
+	/* Task2 */
+	tskid_2 = tk_cre_tsk(&ctsk_2);
+	tk_sta_tsk(tskid_2, 0);
 
 	return 0;
 }
 
-/* Main function for Part2 */
-/*
-int main(void)
-{
-	tm_com_init();
-	tm_putstring("Hello World!!\n");
 
-    while(1){
-        out_w(GPIO_OUT_XOR, (1<<25));	// Switch LED light
-        delay_ms(500);                  // Wait 500msec
-    }
 
-    return 0;
-}
-*/
