+++ include/trykernel.h
@@ -1,4 +1,7 @@
 ﻿#include "typedef.h"
 #include "sysdef.h"
 #include "syslib.h"
+#include "config.h"
+#include "error.h"
 #include "knldef.h"
+#include "apidef.h"
