+++ kernel/inittsk.c
@@ -10,11 +10,11 @@ UW	tskstk_ini[256/sizeof(UW)];				// stack for init task
 ID	tskid_ini;								// ID for init task
 
 T_CTSK ctsk_ini = {
-		.tskadr		= TA_HLNG | TA_RNG0 | TA_USERBUF,
+		.tskatr		= TA_HLNG | TA_RNG0 | TA_USERBUF,
 		.task		= initsk,
 		.itskpri	= 1,
 		.stksz		= sizeof(tskstk_ini),
-		.bugptr		= tskstk_ini,
+		.bufptr		= tskstk_ini,
 };
 
 void initsk(INT stacd, void *exinf)
